SURVIVAL: A reference class for modelling survival of cases in a process model.

Reference class containing some properties and methods required for survival modelling.

Fields

data

list A list of tables. Table names can be: el (event log), cel (current eventlog), cp (case profile), ccp (current case profile), cvp (case-attribute profile), ccvp(current case-attribute profile), cpt (case profile time), ccpt (current case profile time), cpag (case profile aggregated), ccpag (current case profile aggregated)

report

list A list of tables containing analysis reports.

settings

list A list of class settings. Has following parameters: caseID_col, eventTime_col, eventType_col, attribute_col, value_col, caseStart_tag, caseEnd_tag, age_varname, deathReason_varname, cp_attributes, normalize_hazard, cp_aggregators, cp_binners, cp_categoricals, cp_attributes

Methods

feed.caseProfile(dataset)

Feed case profile. Argument 'dataset' must be a sparklyr table. Column names must have been defined in settings

feed.eventlog(dataset)

Feed eventlog. Argument 'dataset' must be a sparklyr table. Column names must have been defined in settings

get.caseProfile(full = F)

Returns current values of case features

get.caseProfileTime(full = T)

Returns timing properties of each case

get.hazard(reasons = NULL, categoricals)

This function calculates hazard and survival probability from aggregated observations

goto(time)

Takes you to a given time

plot.death( reasons = NULL, categoricals = NULL, plotter = "plotly", gain = 100, t0 = 0 )

Plots Kaplan-Meier death chart

plot.hazard( reasons = NULL, categoricals = NULL, plotter = "plotly", smooth_k = 6, gain = 100 )

Plots Kaplan-Meier hazard chart

plot.survival( reasons = NULL, categoricals = NULL, plotter = "plotly", gain = 100, t0 = 0 )

Plots Kaplan-Meier survival chart

reset(...)

Reset given table names in data list

set(...)

Set a setting parameter to a value