TRANSYS: A reference class for modelling Transition Systems

Reference class containing some properties and methods required for analysing, modelling, simulating and visualising a Transition Sytem. A Transistopn System is the more general case of a Markov Chain model descibing a system which can change status over time.

Fields

history

data.frame holding history data of status transitions

report

list list of data.frames containing outputs of various query reports. These tables could be:

nodes: data.frame holding nodes of a transition system network. Nodes are equivalent to statuses. links: data.frame holding links of a transition system network. Links(edges) are equivalent to transitions.

Methods

feed.eventlog(dataset, caseID_col = "caseID", status_col = "status", startTime_col = "startTime", caseStartFlag_col = NULL, caseEndFlag_col = NULL, caseStartTags = NULL, caseEndTags = NULL, sort_startTime = T, add_start = T, remove_sst = F, extra_col = NULL)

Feeds an eventlog as a data.frame or tibble to train the transition system model.

get.case.path()

Returns trace, path, DNA or status sequence of each case

get.case.status.duration()

Returns a table containing duration of each case on each status

get.case.status.time()

Returns entry time of each case to each status in a matrix

get.longest_path()

Returns the longest sequence observed

get.nodes(full = F)

Returns the status profile which is a data.frame containing features associated with the statuses of the transition system.
Each status stablish a node in the process map graph from which you can build a Markov-Chain (MC) model and use it to determine the steady state probabilities or run a memory-less process simulation.
If you run this method once, the output table will be accessible via reports$nodes. If the function is called for the second time, it returns the same table as before if you do not reset the object.
Arguments:
* full (logical): If set to true, case filtering is ignored.
Output:
(data.frame): Includes the following columns (columns with keyword 'case' are generated only if settings$include_case_measures is TRUE):
* totExitFreq: Over all cases, total count of transitions from the status to other statuses.
* totEntryFreq: Over all cases, total count of transitions into the status from other statuses .
totalDuration: Sum of duration(time) cases were in the status.
meanDuration: Average duration(time) each case were in the status.
medDuration: Median of duration(time) cases were in the status.
sdDuration: Standard deviation of duration(time) cases were in the status.
nCaseEntry: How many cases have at least once entered into the status.
nCaseExit: How many cases have at least once exited from the status.
meanExitCaseFrequency: indicates how many times in average, each case has exited from the status

get.status.case.time()

Returns entry time to each status for each case in a matrix (transpose of 'case.status.time')